bugfixes; core code changes; complete cycle for mode tests completed
tests now run without errors when user is root. This requires the chattr
program for providing permission denied scenarios.

added error codes for certops.py

api_read_delete: ALNITAK_CERT_DATA environment parameter changed to
ALNITAK_DELETE_CERT_DATA.

added taint to the deploy mode code when errors are recorded. Then errors
that should not taint are already handled.

changed the default value for the 'time' attribute in the state object from
'0' to 'None'.

certops::get_pem: cert errors now added as a warning.

certops::get_cert_data: takes the state object as an argument so that the
record can be used in error messages. Removed exceptions that will never
likely be thrown.

fsops::cleanup_prev_state: added check for 'is_up' when inspecting the
'published' attribute of the previous state so that delete records are not
created when they should not be.

fsops::process_deletes: removed bug about the dictionary being modified
whilst still in the loop for the delete records.